Restructure the state model of the app
Restructured the model such that the main BooksApp component contains
the users books that would be seen in the bookshelf. These books can be
put into different shelves therefore the updateBookShelf function will
act on this state variable/array of books which will cascade to child
components. With the new state model, each child and grandchild
component will need a way to
(1) see the users books
(2) be able to update the shelf or pass that update function down to the
next child
(3) both of the above situations